requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
